Minor improvement

+++ src/staff.cpp
@@ -300,7 +300,7 @@ void Staff::AdjustLedgerLines(ArrayOfLedgerLines &lines, int extension, int minE
     // Now we transfer the adjustments from the inner dash line to the outer dash lines.
     // This ensures that all dashes on the same note/chord obtain the same ledger line extension.
     for (const Adjustment &adjustment : adjustments) {
-        for (int index = 1; index < lines.size(); ++index) {
+        for (size_t index = 1; index < lines.size(); ++index) {
             LedgerLine &outerLine = lines.at(index);
             IterType iterDash = std::find_if(
                 outerLine.m_dashes.begin(), outerLine.m_dashes.end(), [&adjustment](const DashType &dash) {
